Added last commits from master branch: Include MKS TFTs in build test clean up move.c & default config disable AUTO_SAVE_LOAD_LEVELING_VALUE delete status screen in classic ui, keep it simple display long file name from onboard SD file in ICON mode improve BAUDRATE settings, cleanup BAUDRATE label Enable fullscreen Marlin Mode by default Add editorconfig file, fix encoding, translate comments to English clean up labels, translating chinese fix settingPage array out of bound bug, cause some crash Update language_ru.h improve smart filament detect, auto get E steps by M92 after start up fix M104/M140/M109/M190 'B' 'F', etc... parameters bug Fix some Cyrilic letters at "word_unicode.fon" Cleanup readme Update language_ru.h improve power loss recovery, saved when Z has changed, not 1 sec fix temperature sync bugs, TODO: clean up, it's too redundant now fix M906 TMC Current Update language_ru.h add M300 support for buzzer OS timer accuracy changed from 10ms to 1ms Improve UI previews in readme Add Hungarian Language improve rename of firmware/folder to be update to prevent duplicate u… support more file extension, *.g, *.gco, *.gcode

+++ TFT/src/User/Menu/Selectmode.h
@@ -14,13 +14,13 @@ typedef enum
   MKEY_IDLE = IDLE_TOUCH,
 }MKEY_VALUES;
 
-bool LCD_ReadPen(uint8_t intervals);
+bool LCD_ReadPen(uint16_t intervals);
 uint8_t LCD_ReadTouch(void);
 void Touch_Sw(uint8_t num);
 #if defined(MKS_32_V1_4) || defined(MKS_32_V1_3) || defined(MKS_32_V1_2) || defined(MKS_32_V1_1)
-u8 LCD_ButtonTouch(uint8_t intervals);
+u8 LCD_ButtonTouch(uint16_t intervals);
 #else
-bool LCD_BtnTouch(uint8_t intervals);
+bool LCD_BtnTouch(uint16_t intervals);
 #endif
 
 //select mode fun
